+++ src/pages/DetailRecipePage.jsx
@@ -1,11 +1,7 @@
 import Header from '../components/Header'
 import styles from '../styles/RecipeDetailPage/DetailRecipePage.module.css'
-
 import RecipeDetail from '../components/RecipeDetailPageComponents/RecipeDetailSection/RecipeDetail'
 import RecipeReview from '../components/RecipeDetailPageComponents/RecipeReviewSection/RecipeReview'
-
-
-
 
 const DetailRecipePage = () => {
   return (
