+++ src/styles/Button.module.css
@@ -1,13 +1,19 @@
-.button1 {
+button {
   border-radius: 8px;
   font-family: Noto Sans JP;
   font-weight: 600;
   color: white;
   text-decoration: none;
+  transition: background-color 0.5s, color 0.5s;
+  padding: 10px 0px;
+  border-width: 2px;
+  border-style: solid;
+  cursor: pointer;
+  min-width: 150px;
+}
+
+.button1 {
   background-color: #d18d0f;
-  transition: background-color 0.2s;
-  transition: color 0.2s;
-  padding: 10px 70px;
   border: 2px solid #d18d0f;
 }
 
@@ -17,15 +23,7 @@
 }
 
 .button2 {
-  border-radius: 8px;
-  font-family: Noto Sans JP;
-  font-weight: 600;
-  color: white;
-  text-decoration: none;
   background-color: #3d233a;
-  transition: background-color 0.2s;
-  transition: color 0.2s;
-  padding: 10px 70px;
   border: 2px solid #3d233a;
 }
