--- src/styles/ReusableStyle.module.css
@@ -1,26 +1,0 @@
-
-
-:root {
-    --primary-font: bold 16px normal Ubuntu;
-    --black-purple: rgba(61, 35, 58, 1);
-    --light-orange: rgba(255, 196, 131, 1);
-}
-
-
-
-.flexRow {
-    display: flex;
-    flex-direction: row;
-}
-
-.flexColumn {
-    display: flex;
-    flex-direction: column;
-}
-
-/* Flex items at the center of both X,Y axis */
-.flexItemCenter {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-}
