+++ src/styles/globalStyles.css
@@ -1,6 +1,8 @@
 @import url('https://fonts.googleapis.com/css2?family=Noto+Sans+JP:wght@100;300;400;500;700;900&display=swap');
 
-*, *::before, *::after {
+*,
+*::before,
+*::after {
   margin: 0;
   padding: 0;
   box-sizing: border-box;
@@ -9,17 +11,21 @@
 :root {
   font-family: 'Noto Sans JP', sans-serif;
 
-  font-family: Inter, system-ui, Avenir, Helvetica, Arial, sans-serif;
+  /* font-family: Inter, system-ui, Avenir, Helvetica, Arial, sans-serif; */
   line-height: 1.5;
   font-weight: 400;
 
+  --primary-font: 16px normal;
+  --black-purple: #3d233a;
+  --light-orange: #d18d0f;
+
   color: rgba(0, 0, 0, 0.87);
-  background-color: #F1F1F4;
+  background-color: #f1f1f4;
 }
 
 a {
   color: inherit;
-    text-decoration: none;
+  text-decoration: none;
 }
 
 body {
@@ -27,8 +33,29 @@ body {
   /* overflow: hidden; */
 }
 
-h1, h2, h3, h4, h5, h6 {
+h1,
+h2,
+h3,
+h4,
+h5,
+h6 {
   margin: 0;
   padding: 0;
 }
 
+.flexRow {
+  display: flex;
+  flex-direction: row;
+}
+
+.flexColumn {
+  display: flex;
+  flex-direction: column;
+}
+
+/* Flex items at the center of both X,Y axis */
+.flexItemCenter {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
